some refacto start
